+++ src/main/java/wooteco/chess/domain/game/repository/GameRepository.java
@@ -1,4 +1,4 @@
-package wooteco.chess.repository.jdbc;
+package wooteco.chess.domain.game.repository;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 
 public interface GameRepository extends CrudRepository<GameEntity, String> {
-	@Query("SELECT id FROM game")
+	@Query("SELECT id FROM game_entity")
 	List<String> findAllIds();
 
 }
